+++ src/app/patient/patient-fields/patient-fields.component.ts
@@ -2,24 +2,33 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {FieldService} from "../../services/field.service";
 import {Field} from "../../model/field";
 import _moment from "moment";
-import { ControlContainer, FormControl, FormGroupDirective, NgForm, NgModel, ValidationErrors, FormsModule } from "@angular/forms";
+import {
+  ControlContainer,
+  FormControl,
+  FormGroupDirective,
+  FormsModule,
+  NgForm,
+  NgModel,
+  ValidationErrors
+} from "@angular/forms";
 import {ErrorStateMatcher} from "@angular/material/core";
-import { TranslateService, TranslatePipe } from '@ngx-translate/core';
-import { NgFor, NgSwitch, NgIf, NgStyle, NgSwitchCase, NgSwitchDefault } from '@angular/common';
-import { MatIcon } from '@angular/material/icon';
-import { MatFormField, MatLabel, MatSuffix, MatError } from '@angular/material/form-field';
-import { MatInput } from '@angular/material/input';
-import { MatDatepickerInput, MatDatepickerToggle, MatDatepicker } from '@angular/material/datepicker';
-import { MatIconButton } from '@angular/material/button';
-import { CdkCopyToClipboard } from '@angular/cdk/clipboard';
-import { MatSelect, MatOption } from '@angular/material/select';
+import {TranslatePipe, TranslateService} from '@ngx-translate/core';
+import {NgFor, NgIf, NgStyle, NgSwitch, NgSwitchCase, NgSwitchDefault} from '@angular/common';
+import {MatIcon} from '@angular/material/icon';
+import {MatError, MatFormField, MatLabel, MatSuffix} from '@angular/material/form-field';
+import {MatInput} from '@angular/material/input';
+import {MatDatepicker, MatDatepickerInput, MatDatepickerToggle} from '@angular/material/datepicker';
+import {MatIconButton} from '@angular/material/button';
+import {CdkCopyToClipboard} from '@angular/cdk/clipboard';
+import {MatOption, MatSelect} from '@angular/material/select';
+import {InvalidBirthdayDirective} from "../../shared/directives/invalid-birthday-directive";
 
 @Component({
     selector: 'app-patient-fields',
     templateUrl: './patient-fields.component.html',
     styleUrls: ['./patient-fields.component.css'],
     viewProviders: [{ provide: ControlContainer, useExisting: NgForm }],
-    imports: [FormsModule, NgFor, NgSwitch, NgIf, NgStyle, MatIcon, NgSwitchCase, MatFormField, MatLabel, MatInput, MatDatepickerInput, MatDatepickerToggle, MatSuffix, MatDatepicker, MatIconButton, CdkCopyToClipboard, MatSelect, MatOption, NgSwitchDefault, MatError, TranslatePipe]
+  imports: [FormsModule, NgFor, NgSwitch, NgIf, NgStyle, MatIcon, NgSwitchCase, MatFormField, MatLabel, MatInput, MatDatepickerInput, MatDatepickerToggle, MatSuffix, MatDatepicker, MatIconButton, CdkCopyToClipboard, MatSelect, MatOption, NgSwitchDefault, MatError, TranslatePipe, InvalidBirthdayDirective]
 })
 
 export class PatientFieldsComponent implements OnInit {
@@ -57,8 +66,10 @@ export class PatientFieldsComponent implements OnInit {
       return this.translate.instant('patientFields.error_value_text1') + " \"" + fieldName + "\" " + this.translate.instant('patientFields.error_value_text2');
     else if (errors?.['required'])
       return this.translate.instant('patientFields.error_mandatory_text1') + " \"" + fieldName + "\" " +  this.translate.instant('patientFields.error_mandatory_text2');
+    else if (errors?.['invalidBirthday'])
+      return this.translate.instant('patientFields.invalid_birthday');
     else
-      return "fehler";
+      return "invalid";
   }
 
   enableDateFieldValidation(field: NgModel) {
@@ -77,7 +88,7 @@ export class PatientFieldsComponent implements OnInit {
       //let isFieldsEmpty = !Object.keys(this.fields).length || !Object.entries(this.fields).some( (e) => e[1].length > 0);
       return field.invalid &&
           (field.dirty || field.touched) &&
-          (field.errors?.['pattern'] || field.errors?.['required']);
+          (field.errors?.['pattern'] || field.errors?.['required'] || field.errors?.['invalidBirthday']);
     }
   }